reglages

+++ app/Http/Controllers/HotelController.php
@@ -8,7 +8,9 @@ use Illuminate\Support\Facades\Auth;
 class HotelController extends Controller
 {
     public function index(){
-        return response()->json(Hotel::all());
+      return response()->json(
+        Hotel::where('user_id', Auth::id())->get()
+    );
     }
 
     public function store(Request $request){
@@ -20,6 +22,7 @@ $validated=$request->validate([
 'devise'=>'required|string',
 'prixNuitee'=>'required|numeric|min:0',
 "cheminImage" => "required|file|image|max:2048",
+
 
 
 
@@ -38,9 +41,9 @@ return response()->json($hotel);
     public function update(Request $request, $id) {
         $hotel = Hotel::findOrFail($id);
 
-        // if ($hotel->user_id !== Auth::id()) {
-        //     return response()->json(['message' => 'Unauthorized'], 403);
-        // }
+        if ($hotel->user_id !== Auth::id()) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
 
        $validated=$request->validate([
            'nomHotel'=>'required|string',
@@ -65,9 +68,9 @@ $hotel->update($validated);
     {
         $hotel = Hotel::findOrFail($id);
 
-        // if ($hotel->user_id !== Auth::id()) {
-        //     return response()->json(['message' => 'Non autorisé'], 403);
-        // }
+        if ($hotel->user_id !== Auth::id()) {
+            return response()->json(['message' => 'Non autorisé'], 403);
+        }
 
         $hotel->delete();
